Fixes change request in PR #84
